Changed how text was drawn to draw text using System.Drawing onto a bitmap and then load that bitmap onto a textured quad as opposed to using a signed distance field + bitmap font.
Converted instances where TextComponent was being used to use the new Text class.

Changed the generic instanced render function to not spazz out when dealing with objects containing multiple base objects

Moved unit status bars into a scene controlled parent object instead of adding them directly to the scene's UI

Added the ability to dispose of loaded textures.

Moved most of the handling that was occurring in the engine tools over here and added the engine as a dependency to the engine tools. That way changes can be made in the engine and propagated out to where they need to be instead of having to maintain parity between the two.

Added support for tile overlays to the tiletexturer although I am currently working on a feature that will hopefully make the tiletexturer unnecessary

Started the process of adding a quest log but got sidetracked. Will continue at some point.

Changed how unit parameters were  being applied to use reflection. If a parameter matches a field of the Unit or UnitInfo class then the value of that parameter will be applied to that field

Made more texture slots available to the main shader.

Increased the high frequency ticks per second from ~80 to ~120

+++ MortalDungeon/Game/Serializers/Animations/Animation.cs
@@ -39,7 +39,7 @@ namespace MortalDungeon.Game.Serializers
             foreach(int frameIndex in FrameIndices)
             {
                 RenderableObject frame = new RenderableObject(
-                    new SpritesheetObject(frameIndex, Spritesheets.AllSpritesheets[Spritesheet]).CreateObjectDefinition(true), 
+                    new SpritesheetObject(frameIndex, Spritesheets.AllSpritesheets[(int)Spritesheet]).CreateObjectDefinition(true), 
                     WindowConstants.FullColor, ObjectRenderType.Texture, Shaders.DEFAULT_SHADER);
 
                 builtAnim.Frames.Add(frame);
